+++ src/main/java/es/grupoO/FastFood/services/RepartidoresService.java
@@ -1,5 +1,10 @@
 package es.grupoO.FastFood.services;
+
+import es.grupoO.FastFood.exceptions.NoExistDBException;
+import es.grupoO.FastFood.exceptions.UsernameAlreadyExistException;
+import es.grupoO.FastFood.exceptions.NoMatchingPasswordException;
 import es.grupoO.FastFood.factory.RepartidorFactory;
+import es.grupoO.FastFood.model.valueobject.Pair;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -8,31 +13,32 @@ import es.grupoO.FastFood.repository.RepartidoresRepository;
 import es.grupoO.FastFood.model.valueobject.Email;
 import es.grupoO.FastFood.auth.HashMaker;
 
-
 @Service
 public class RepartidoresService {
     @Autowired
     private RepartidoresRepository repository;
+    @Autowired
+    private AuthService authService;
 
     public Repartidor validar(String email, String passwd) {
-        // TODO: puede petar 
-        Repartidor repartidor = this.repository.findByEmail(Email.parse(email));
-        HashMaker hasher = new HashMaker();
-        if(!hasher.verify(repartidor.getHashPassword(), passwd)) {
-         // TODO: verificar   
-         return null;
-        } else {
-            return repartidor;
-        }
+        Pair<Repartidor, String> data = this.authService.loginRepartidor(email, passwd);
+        // TODO: Token
+        return data.getFirst();
     }
 
-    public Repartidor buscarRepartidorPorID(ObjectId idRepar){
-        // TODO: Puede petar
-        return this.repository.findById(idRepar).get();
+    public Repartidor buscarRepartidorPorID(ObjectId idRepar) {
+        Repartidor repartidor = this.repository.findById(idRepar).orElse(null);
+        if (repartidor == null) {
+            throw new NoExistDBException("El repartidor no esta registrado");
+        }
+        return repartidor;
     }
 
     public void insertarRepartidor(String nombre, String telefono, String email, String passwd) {
-        //TODO comprobar condicione de insertar
+        if(this.repository.findByEmail(Email.parse(email)) == null) {
+            throw new UsernameAlreadyExistException("El repartidor ya existe");
+        }
+        
         RepartidorFactory repartidorFactory = new RepartidorFactory(nombre, telefono, email, passwd);
         Repartidor repartidor = repartidorFactory.fabricarRepartidor();
         
@@ -40,15 +46,21 @@ public class RepartidoresService {
     }
 
     public void borrarRepartidor(ObjectId idRepar) {
-        //TODO puede petar
+        Repartidor repartidor = this.repository.findById(idRepar).orElse(null);
+        if (repartidor == null) {
+            throw new NoExistDBException("El repartidor no esta registrado");
+        }
         this.repository.deleteById(idRepar);
     }
 
     public void passwdChanger(ObjectId idRepartidor, String newPasswd) {
-        //TODO puede petar
         Repartidor repartidor = this.buscarRepartidorPorID(idRepartidor);
+        if(repartidor == null) {
+            throw new NoExistDBException("El repartidor no esta registrado");
+        }
         HashMaker hasher = new HashMaker();
         repartidor.setHashPassword(hasher.encoder(newPasswd));
 
-        this.repository.save(repartidor);    }
+        this.repository.save(repartidor);
+    }
 }
