+++ src/main/java/es/grupoO/FastFood/services/RestaurantesService.java
@@ -3,6 +3,7 @@ import es.grupoO.FastFood.factory.RestauranteFactory;
 import es.grupoO.FastFood.model.entity.Restaurante;
 import es.grupoO.FastFood.model.entity.Valoracion;
 import es.grupoO.FastFood.model.valueobject.Email;
+import es.grupoO.FastFood.model.valueobject.Pair;
 import es.grupoO.FastFood.repository.RestaurantesRepository;
 import es.grupoO.FastFood.repository.ValoracionesRepository;
 import org.bson.types.ObjectId;
@@ -11,6 +12,9 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import es.grupoO.FastFood.auth.HashMaker;
 
+import es.grupoO.FastFood.exceptions.NoExistDBException;
+import es.grupoO.FastFood.exceptions.UsernameAlreadyExistException;
+import es.grupoO.FastFood.exceptions.NoMatchingPasswordException;
 
 @Service
 public class RestaurantesService {
@@ -18,36 +22,34 @@ public class RestaurantesService {
     private RestaurantesRepository repository;
     @Autowired
     private ValoracionesRepository valoracionesRepository;
+    @Autowired
+    private AuthService authService;
 
     public Restaurante validar(String email, String passwd) {
-        //TODO puede petar
-        Restaurante rest = this.repository.findByEmail(Email.parse(email));
-        if(rest == null) {
-            // TODO
-            return null;
-        }
+        Pair<Restaurante, String> data = this.authService.loginRestaurante(email, passwd);
+        // TODO: Token
+        return data.getFirst();
+    }
 
-        HashMaker hasher = new HashMaker();
-        if(!hasher.verify(rest.gethashPassword(), passwd)) {
-            //TODO   
-            return null;
+    public Restaurante buscarRestaurantePorID(ObjectId idRest) {
+        Restaurante rest = this.repository.findById(idRest).orElse(null);
+        if(rest == null) {
+            throw new NoExistDBException("El restaurante no existe");
         }
         return rest;
     }
 
-    public Restaurante buscarRestaurantePorID(ObjectId idRest) {
-        //TODO puede petar
-        return this.repository.findById(idRest).get();
-    }
-
     public List<Restaurante> buscarRestaurante(String nombre) {
-        //TODO puede petar
-        return this.repository.findAllByNombreContaining(nombre);
+        List<Restaurante> restaurantes = this.repository.findAllByNombreContaining(nombre);
+        if(restaurantes == null) {
+            throw new NoExistDBException("El restaurante no existe");
+        }
+        return restaurantes;
     }
 
     public Restaurante insertarRestaurante(String nombre, int categoria, String direccion,String telefono, 
         String email, String horaApertura, String horaCierre, String passwd){
-            //TODO comprobar condiciones de insercion
+        //TODO comprobar condiciones de insercion
         RestauranteFactory fact = new RestauranteFactory(nombre, direccion, telefono, horaApertura, horaCierre, categoria, email, passwd);
         Restaurante restaurante = fact.fabricarRestaurante();
         Valoracion val = restaurante.getValoracion();
@@ -57,28 +59,31 @@ public class RestaurantesService {
     }
 
     public void borrarRestaurante(ObjectId id) {
-        //TODO puede petar
+        Restaurante restaurante = this.buscarRestaurantePorID(id);
+        if(restaurante == null) {
+            throw new NoExistDBException("El restaurante no existe");
+        }
         this.repository.deleteById(id);
     }
 
     public void actualizarValoracion(ObjectId id, int valor) {
-        //TODO puede petar 
         Restaurante restaurante = this.buscarRestaurantePorID(id);
         if(restaurante == null) {
-            // TODO: Error
+            throw new NoExistDBException("El restaurante no existe");
         }
         
         Valoracion valoracion = restaurante.getValoracion();
         valoracion.actualizarValoracion(valor);
         
         this.valoracionesRepository.save(valoracion);
-        
     }
 
     public void passwdChanger(ObjectId idRest, String newPasswd) {
         //TODO comprobar que el usuario cambia su propia contraseña
-        //TODO puede petar si el rest no existe
         Restaurante rest = this.buscarRestaurantePorID(idRest);
+        if(rest == null) {
+            throw new NoExistDBException("El restaurante no existe");
+        }
         HashMaker hasher = new HashMaker();
         rest.setHashPassword(hasher.encoder(newPasswd));
